+++ src/geo/projection/projection_factory.ts
@@ -19,7 +19,7 @@ export function createProjectionFromName(name: ProjectionSpecification['type'], 
     transform: ITransform;
     cameraHelper: ICameraHelper;
 } {
-    const transformOptions = {constrain: transformConstrain};
+    const transformOptions = {constrainOverride: transformConstrain};
     if (Array.isArray(name)) {
         const globeProjection = new GlobeProjection({type: name});
         return {